backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
alembic
httpx
python-dotenv
pydantic
openai
python-jose[cryptography]
passlib[bcrypt]
pytest
pytest-asyncio
aiosqlite
aiomysql
aiohttp


# cd backend
# C:\Users\Asus\AppData\Local\Programs\Python\Python311\python.exe -m pip install -r requirements.txt
